Fixes interface loop on Panels communication
