+++ src/org/hopandfork/jgnuplot/gui/panel/BottomInterface.java
@@ -1,0 +1,17 @@
+package org.hopandfork.jgnuplot.gui.panel;
+
+import javax.swing.JButton;
+
+public interface BottomInterface {
+		public JButton getbEdit();
+
+		public JButton getbDelete();
+
+		public JButton getbAdd();
+
+		public JButton getbMoveUp();
+
+		public JButton getbMoveDown();
+		
+		public void reset();
+}
