+++ src/org/hopandfork/jgnuplot/gui/panel/MainInterface.java
@@ -1,0 +1,6 @@
+package org.hopandfork.jgnuplot.gui.panel;
+
+public interface MainInterface {
+	public void setTitle(String title);
+	public void reset();
+}
