+++ src/org/hopandfork/jgnuplot/gui/panel/MenuInterface.java
@@ -1,0 +1,13 @@
+package org.hopandfork.jgnuplot.gui.panel;
+
+import javax.swing.JMenuItem;
+
+public interface MenuInterface {
+		public JMenuItem getEdit_menu_item();
+
+		public JMenuItem getDelete_menu_item();
+
+		public JMenuItem getMoveup_menu_item();
+
+		public JMenuItem getMovedown_menu_item();
+}
