+++ src/org/hopandfork/jgnuplot/utility/UpdateChecker.java
@@ -1,17 +1,14 @@
 package org.hopandfork.jgnuplot.utility;
 
 import java.io.File;
-import java.io.IOException;
 import java.util.List;
 
-import org.hopandfork.jgnuplot.JGP;
 import org.hopandfork.jgnuplot.control.PlottableDataController;
 import org.hopandfork.jgnuplot.model.DataFile;
 import org.hopandfork.jgnuplot.model.PlottableData;
 
 @Deprecated
 public class UpdateChecker implements Runnable {
-	private JGP owner;
 
 	public boolean isCheckForUpdate() {
 		return checkForUpdate;
@@ -24,8 +21,7 @@ public class UpdateChecker implements Runnable {
 	private boolean checkForUpdate;
 	private PlottableDataController plottableDataController;
 
-	public UpdateChecker(JGP owner, PlottableDataController plottableDataController) {
-		this.owner = owner;
+	public UpdateChecker( PlottableDataController plottableDataController) {
 		this.plottableDataController = plottableDataController;
 	}
